refactor: split builtin lowering facade
